+++ ios/Classes/SwiftFlutterBranchSdkPlugin.swift
@@ -209,6 +209,18 @@ public class SwiftFlutterBranchSdkPlugin: NSObject, FlutterPlugin, FlutterStream
         case "getAdvertisingIdentifier" :
             getAdvertisingIdentifier(result: result)
             break
+        case "setConnectTimeout":
+            setConnectTimeout(call: call)
+            break
+        case "setRetryCount":
+            setRetryCount(call: call)
+            break
+        case "setRetryInterval":
+            setRetryInterval(call: call)
+            break
+        case "setTimeout":
+            setTimeout(call: call)
+            break
         default:
             result(FlutterMethodNotImplemented)
         }
@@ -254,7 +266,6 @@ public class SwiftFlutterBranchSdkPlugin: NSObject, FlutterPlugin, FlutterStream
         
         let response : NSMutableDictionary! = [:]
         buo?.showShareSheet(with: lp, andShareText: shareText, from: controller) { (activityType, completed, error) in
-            print(activityType ?? "")
             if completed {
                 response["success"] = NSNumber(value: true)
             } else {
@@ -278,11 +289,13 @@ public class SwiftFlutterBranchSdkPlugin: NSObject, FlutterPlugin, FlutterStream
     
     private func trackContent(call: FlutterMethodCall) {
         let args = call.arguments as! [String: Any?]
-        let buoDict = args["buo"] as! [String: Any?]
+        let buoDict = args["buo"] as! [[String: Any?]]
         let eventDict = args["event"] as! [String: Any?]
-        let buo: BranchUniversalObject? = convertToBUO(dict: buoDict)
+        let buoList: [BranchUniversalObject] = buoDict.map { b in
+            convertToBUO(dict: b)!
+        }
         let event: BranchEvent? = convertToEvent(dict : eventDict)
-        event!.contentItems = [ buo! ]
+        event!.contentItems =  buoList
         
         DispatchQueue.main.async {
             event!.logEvent()
@@ -583,6 +596,31 @@ public class SwiftFlutterBranchSdkPlugin: NSObject, FlutterPlugin, FlutterStream
             DispatchQueue.main.async {
                 result(String(""))  // return notSupported
             }
+        }
+    }
+    private func setTimeout(call: FlutterMethodCall) {
+        let args = call.arguments as! [String: Any?]
+        let timeout = args["timeout"] as? Int ?? 0
+    }
+
+    private func setConnectTimeout(call: FlutterMethodCall) {
+        let args = call.arguments as! [String: Any?]
+        let connectTimeout = args["connectTimeout"] as? Int ?? 0
+        DispatchQueue.main.async {
+            Branch.getInstance().setNetworkTimeout(TimeInterval(connectTimeout))
+        }
+    }
+    
+    private func setRetryCount(call: FlutterMethodCall) {
+        let args = call.arguments as! [String: Any?]
+        let retryCount = args["retryCount"] as? Int ?? 0
+    }
+
+    private func setRetryInterval(call: FlutterMethodCall) {
+        let args = call.arguments as! [String: Any?]
+        let retryInterval = args["retryInterval"] as? Int ?? 0
+        DispatchQueue.main.async {
+            Branch.getInstance().setRetryInterval(TimeInterval(retryInterval))
         }
     }
 }