+++ lib/src/flutter_branch_sdk_mobile.dart
@@ -30,6 +30,7 @@ class FlutterBranchSdkMobile implements FlutterBranchSdkPlatform {
 
   ///Identifies the current user to the Branch API by supplying a unique identifier as a userId value
 
+  @Deprecated('version 5.0.0')
   @override
   void initWeb({required String branchKey}) {
     //nothing
@@ -149,10 +150,11 @@ class FlutterBranchSdkMobile implements FlutterBranchSdkPlatform {
   ///Logs this BranchEvent to Branch for tracking and analytics
   @override
   void trackContent(
-      {required BranchUniversalObject buo, required BranchEvent branchEvent}) {
+      {required List<BranchUniversalObject> buo,
+      required BranchEvent branchEvent}) {
     Map<String, dynamic> _params = {};
 
-    _params['buo'] = buo.toMap();
+    _params['buo'] = buo.map((b) => b.toMap()).toList();
     if (branchEvent.toMap().isNotEmpty) {
       _params['event'] = branchEvent.toMap();
     }
@@ -329,5 +331,33 @@ class FlutterBranchSdkMobile implements FlutterBranchSdkPlatform {
     final String uuid = (await _messageChannel
         .invokeMethod<String>('getAdvertisingIdentifier'))!;
     return uuid;
+  }
+
+  @override
+  void setConnectTimeout(int connectTimeout) {
+    Map<String, dynamic> _params = {};
+    _params['connectTimeout'] = connectTimeout;
+    _messageChannel.invokeMethod('setConnectTimeout', _params);
+  }
+
+  @override
+  void setRetryCount(int retryCount) {
+    Map<String, dynamic> _params = {};
+    _params['retryCount'] = retryCount;
+    _messageChannel.invokeMethod('setRetryCount', _params);
+  }
+
+  @override
+  void setRetryInterval(int retryInterval) {
+    Map<String, dynamic> _params = {};
+    _params['retryInterval'] = retryInterval;
+    _messageChannel.invokeMethod('setRetryInterval', _params);
+  }
+
+  @override
+  void setTimeout(int timeout) {
+    Map<String, dynamic> _params = {};
+    _params['timeout'] = timeout;
+    _messageChannel.invokeMethod('setTimeout', _params);
   }
 }